.github/actions/toc/requirements.txt
python-frontmatter==1.0.0
toml==0.10.2
